Support window filter allowedScreens filter override
Allow the user to set a list of allowed screens for PaperWM to tile,
using the hs.window.filter:setScreens() method. Window events will not
be received for windows that are on an ignored screen. When moving a
window to or from an ignored screen, check to see whether the screen is
allowed before removing, adding, or tiling the window(s) on the screen.